requirements.txt
numpy
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
mlx
Pillow